Transfer scoring logic to separate file
