+++ server/elasticsearch/index.js
@@ -1,4 +1,5 @@
 const client = require("../controllers/elasticsearch.js");
+const scoring = require("./scoring");
 const deleteByQuery = require("elasticsearch-deletebyquery");
 
 const index = "products";
@@ -20,74 +21,6 @@ const addDocument = (body) => client.index({
     console.log(resp);
 });
 
-const foldHits = (hits) => {
-    let folded = [];
-    let result = [];
-
-    hits.forEach((hit) => {
-        folded[hit._source.key] = hit;
-        folded[hit._source.key].count = 0;
-    });
-
-    hits.forEach((hit) => {
-        folded[hit._source.key].count++;
-    });
-
-    for (key in folded) {
-        if (folded.hasOwnProperty(key)) {
-            result.push(folded[key]);
-        }
-    };
-
-    return result;
-};
-
-const calculateScore = (hits) => {
-    const factor = 1;
-    const countCoefficient = 0.01;
-    const relevanceCoefficient = 0;
-
-    let countFactor = 0;
-
-    hits.forEach((hit) => {
-        countFactor = factor + (hit.count * countCoefficient);
-        hit._score = hit._score * countFactor;
-    });
-
-    return hits;
-};
-
-const scoreComparsion = (leftHit, rightHit) => {
-    if (leftHit._score < rightHit._score) {
-        return -1;
-    } else if (leftHit._score > rightHit._score) {
-        return 1;
-    }
-    return 0;
-};
-
-const reduceHits = (hits) => {
-    const amount = 3;
-
-    hits.sort(scoreComparsion)
-    if (hits.length >= amount) {
-        return hits.slice(0, amount);
-    }
-    return hits;
-};
-
-const handleHits = (hits) => {
-    result = calculateScore(foldHits(hits));
-
-    result.forEach(hit => {
-        console.log(hit._source.name);
-        //console.log(hit._source.email);
-        console.log(hit._score)
-    });
-
-    return reduceHits(result);
-};
-
 const getDocuments = (query, response) => client.search({
     index,
     type,
@@ -105,13 +38,7 @@ const getDocuments = (query, response) => client.search({
         }
     }
 }).then((resp) => {
-    resp.hits.hits.forEach(hit => {
-        //console.log(hit._source.name);
-        //console.log(hit._source.email);
-        //console.log(hit._score)
-        //console.log(hit);
-    });
-    response.json(handleHits(resp.hits.hits));
+    response.json(scoring.handleHits(resp.hits.hits));
 });
 
 const getAllDocuments = () => client.search({
